Move all stores in "core/stores"
