+++ libs/search-widget/src/_video-widget/SearchBar.svelte
@@ -2,7 +2,7 @@
 
 <script lang="ts">
   import type { KBStates } from '@nuclia/core';
-  import { resetStore, nucliaStore } from '../core/store';
+  import { resetStore, nucliaStore } from '../core/stores/main.store';
   import { initNuclia, resetNuclia } from '../core/api';
   import { onMount } from 'svelte';
   import { setCDN, loadFonts, loadSvgSprite } from '../core/utils';
