+++ libs/search-widget/src/viewer/InputViewer.svelte
@@ -2,7 +2,7 @@
   import { _ } from '../core/i18n';
   import { getCDN } from '../core/utils';
   import { SearchOrder } from '../core/models';
-  import { viewerStore, viewerState, clearSearch } from './viewer.store';
+  import { viewerStore, viewerState, clearSearch } from '../core/stores/viewer.store';
 
   const query = viewerStore.query;
   query['set'] = query.next;
