+++ libs/search-widget/src/viewer/paragraphs/ParagraphWithAnnotations.svelte
@@ -3,10 +3,10 @@
   import { onDestroy, onMount } from 'svelte';
   import EntityFamilyMenu from '../menus/EntityFamilyMenu.svelte';
   import { EntityGroup, WidgetParagraph } from '../../core/models';
-  import { viewerStore } from '../viewer.store';
+  import { viewerStore } from '../../core/stores/viewer.store';
   import { map } from 'rxjs';
-  import { addAnnotation, Annotation, removeAnnotation, updateAnnotation } from '../stores/annotation.store';
-  import { addEntity, nucliaStore } from '../../core/store';
+  import { addAnnotation, Annotation, removeAnnotation, updateAnnotation } from '../../core/stores/annotation.store';
+  import { addEntity, nucliaStore } from '../../core/stores/main.store';
   import { Duration } from '../../common/transition.utils';
 
   export let paragraph: WidgetParagraph;
