+++ libs/search-widget/src/viewer/paragraphs/ParagraphWithMenu.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import { hasAuthData } from '../../core/api';
-  import { nucliaState } from '../../core/store';
+  import { nucliaState } from '../../core/stores/main.store';
   import { filter, map } from 'rxjs';
   import Paragraph from './Paragraph.svelte';
   import LabelMenu from '../menus/LabelMenu.svelte';
